// src/app/core/interfaces/tour.interfaces.ts
export interface tourDataType{
    id?: Number;
    title?: String;
    img?: String;
    Oldprice?: number;
    Newprice?: number;
    AdditionalInformation?:string[];
    Dateofaddition?: String;
    Location?: String;
    Difficulty?: String;
    totaldistance?: String;
    Views?: number;
    Rate?: String;
    Thetourincludes?: String;
    category?:string;
 }